requirements.txt
blinker==1.8.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configparser==4.0.2
cryptography==43.0.0
dissect.cstruct==4.0
dissect.esedb==3.14
dissect.util==3.17
dnspython==2.6.1
dsinternals==1.2.4
Flask==3.0.3
future==1.0.0
hexdump==3.3
impacket==0.11.0
itsdangerous==2.2.0
Jinja2==3.1.4
ldap3==2.9.1
ldapdomaindump==0.9.4
lxml==5.2.2
MarkupSafe==2.1.5
more-itertools==5.0.0
pyasn1==0.6.0
pycparser==2.22
pycryptodomex==3.20.0
pyOpenSSL==24.2.1
pyparsing==2.4.7
python-evtx==0.7.4
six==1.16.0
Werkzeug==3.0.3
zipp==1.0.0
